# api/v2/openapi.yaml
openapi: 3.0.3
info:
  title: Assisted Migration Agent API
  version: v2
servers:
  - url: /api/v2
...
paths:
...
  /collections/{id}/inventory:
    get:
      tags: [Inventories]
      summary: Get collected inventory for a collection
      operationId: getInventory
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
      responses:
        '200':
          description: Collected inventory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inventory'
        '404':
          description: Collection or inventory not found
        '500':
          description: Internal server error

  # ── Virtual Machines ───────────────────────────────────────────────────
  /collections/{id}/virtualmachines:
    get:
      tags: [VirtualMachines]
      summary: List VirtualMachines in a collection
      operationId: listVirtualMachines
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
        - name: byExpression
          in: query
          description: Filter by expression
          schema:
            type: string
        - name: sort
          in: query
          description: Sort fields with direction (e.g., "name:asc", "cluster:desc"). Valid fields are name, vCenterState, cluster, diskSize, memory, issues, cpuUsage, diskUsage, ramUsage, cpuAvg, memAvg.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
          example: ["cluster:asc", "name:desc"]
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: pageSize
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: List of VirtualMachines
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VirtualMachineListResponse'
        '400':
          description: Invalid request parameters
        '404':
          description: Collection not found
        '500':
          description: Internal server error
...
  /collections/{id}/groups:
    get:
      tags: [Groups]
      summary: List all groups in a collection
      operationId: listGroups
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
        - name: byName
          in: query
          required: false
          description: Filter groups by name (case-insensitive substring match)
          schema:
            type: string
        - name: page
          in: query
          required: false
          description: Page number (1-indexed)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          description: Number of groups per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of groups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupListResponse'
        '400':
          description: Invalid filter expression
        '404':
          description: Collection not found
        '500':
          description: Internal server error
    post:
      tags: [Groups]
      summary: Create a new group in a collection
      operationId: createGroup
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Invalid request
        '404':
          description: Collection not found
        '500':
          description: Internal server error
  /collections/{id}/groups/{groupId}:
    get:
      tags: [Groups]
      summary: Get group by ID with its VirtualMachines
      operationId: getGroup
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
        - name: groupId
          in: path
          required: true
          description: Group ID
          schema:
            type: string
        - name: sort
          in: query
          description: Sort fields with direction (e.g., "name:asc", "cluster:desc"). Valid fields are name, vCenterState, cluster, diskSize, memory, issues, cpuUsage, diskUsage, ramUsage, cpuAvg, memAvg.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
          example: ["cluster:asc", "name:desc"]
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: pageSize
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Group details with VirtualMachines
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '404':
          description: Collection or group not found
        '500':
          description: Internal server error
    patch:
      tags: [Groups]
      summary: Update group
      operationId: updateGroup
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
        - name: groupId
          in: path
          required: true
          description: Group ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupRequest'
      responses:
        '200':
          description: Group updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Invalid request
        '404':
          description: Collection or group not found
        '500':
          description: Internal server error
    delete:
      tags: [Groups]
      summary: Delete group
      operationId: deleteGroup
      parameters:
        - name: id
          in: path
          required: true
          description: Collection ID
          schema:
            type: string
        - name: groupId
          in: path
          required: true
          description: Group ID
          schema:
            type: string
      responses:
        '204':
          description: Group deleted
        '404':
          description: Collection or group not found
        '500':
          description: Internal server error

  # ── Credentials ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Inventory:
      type: object
      required:
        - inventory
      properties:
        inventory:
          $ref: 'https://raw.githubusercontent.com/kubev2v/migration-planner/main/api/v1alpha1/openapi.yaml#/components/schemas/UpdateInventory'

    # ── VCenter Credentials ──────────────────────────────────────────────
...
    VirtualMachine:
      type: object
      required:
        - name
        - id
        - vCenterID
        - vCenterState
        - cluster
        - datacenter
        - diskSize
        - memory
        - issueCount
      properties:
        name:
          type: string
          description: VirtualMachine name
        id:
          type: string
          description: VirtualMachine ID in vCenter
        vCenterID:
          type: string
          description: Identifier of the vCenter this VirtualMachine belongs to
        vCenterState:
          type: string
          description: vCenter state (e.g., poweredOn, poweredOff, suspended)
        cluster:
          type: string
          description: Cluster name
        datacenter:
          type: string
          description: Datacenter name
        diskSize:
          type: integer
          format: int64
          description: Total disk size in MB
        memory:
          type: integer
          format: int64
          description: Memory size in MB
        issueCount:
          type: integer
          description: Number of issues found for this VirtualMachine
        migratable:
          type: boolean
          description: True if the VirtualMachine is migratable for MTV
        template:
          type: boolean
          description: True if the VirtualMachine is a template
        inspectionStatus:
          $ref: '#/components/schemas/InspectionStatus'
        inspectionConcernCount:
          type: integer
          description: Number of inspection concerns for the latest persisted result
        tags:
          type: array
          items:
            type: string
          description: Tags aggregated from matching groups
...
    VirtualMachineListResponse:
      type: object
      required:
        - virtualMachines
        - total
        - page
        - pageCount
      properties:
        virtualMachines:
          type: array
          items:
            $ref: '#/components/schemas/VirtualMachine'
        total:
          type: integer
          description: Total number of VirtualMachines matching the filter
        page:
          type: integer
          description: Current page number
        pageCount:
          type: integer
          description: Total number of pages

    # ── VirtualMachine Sub-resources ─────────────────────────────────────────────────
...
    InspectionStatus:
      type: object
      required:
        - state
      properties:
        state:
          type: string
          enum:
            - pending
            - running
            - completed
            - canceled
            - error
          x-enum-varnames:
            - InspectionStatusStatePending
            - InspectionStatusStateRunning
            - InspectionStatusStateCompleted
            - InspectionStatusStateCanceled
            - InspectionStatusStateError
          description: Current inspection state
        error:
          type: string
          description: Error message when state is error
...
    Group:
      type: object
      required:
        - id
        - name
        - filter
      properties:
        id:
          type: string
          description: Unique group identifier
        name:
          type: string
          description: Group display name
        description:
          type: string
          description: Optional group description
        filter:
          type: string
          description: Filter expression (pkg/filter DSL)
          example: "memory >= 8GB and cluster = 'prod'"
        tags:
          type: array
          items:
            type: string
          description: Tags applied to matching VirtualMachines
          example: ["production", "critical"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GroupListResponse:
      type: object
      required:
        - groups
        - total
        - page
        - pageCount
      properties:
        groups:
          type: array
          items:
            $ref: '#/components/schemas/Group'
        total:
          type: integer
        page:
          type: integer
        pageCount:
          type: integer
    GroupResponse:
      type: object
      required:
        - group
        - vms
        - total
        - page
        - pageCount
      properties:
        group:
          $ref: '#/components/schemas/Group'
        vms:
          type: array
          items:
            $ref: '#/components/schemas/VirtualMachine'
        total:
          type: integer
        page:
          type: integer
        pageCount:
          type: integer
    CreateGroupRequest:
      type: object
      required:
        - name
        - filter
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          x-oapi-codegen-extra-tags:
            binding: "required,min=1,max=100"
        description:
          type: string
          maxLength: 500
          x-oapi-codegen-extra-tags:
            binding: "omitempty,max=500"
        filter:
          type: string
          minLength: 1
          x-oapi-codegen-extra-tags:
            binding: "required,min=1"
        tags:
          type: array
          items:
            type: string
            pattern: "^[a-zA-Z0-9_.]+$"
          x-oapi-codegen-extra-tags:
            binding: "omitempty,dive,tag_format"
    UpdateGroupRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          x-oapi-codegen-extra-tags:
            binding: "omitempty,min=1,max=100"
        description:
          type: string
          maxLength: 500
          x-oapi-codegen-extra-tags:
            binding: "omitempty,max=500"
        filter:
          type: string
          minLength: 1
          x-oapi-codegen-extra-tags:
            binding: "omitempty,min=1"
        tags:
          type: array
          items:
            type: string
            pattern: "^[a-zA-Z0-9_.]+$"
          x-oapi-codegen-extra-tags:
            binding: "omitempty,dive,tag_format"

    # ── Applications ─────────────────────────────────────────────────────
...